LayoutTests/imported/w3c/web-platform-tests/dom/nodes/moveBefore/mutation-observer-expected.txt


FAIL [Connected move] MutationObserver removal + insertion is tracked by moveBefore() assert_equals: MutationObserver has emitted a single mutation list expected 1 but got 0
FAIL [Disconnected move] MutationObserver removal + insertion is tracked by moveBefore() assert_equals: MutationObserver has emitted a single mutation list expected 1 but got 0
